requirements.txt
fastapi
uvicorn
opencv-python
pyaudio
google-generativeai
openai
requests
python-dotenv
cartesia